# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte Allotjaments Sol",
        "description": "REST API. Projecte Allotjaments Sol. DAW Client i servidor. Per utilitzar les cridades que necesiten token primer\n  s'haura de fer login amb un usuari per obtenir el token i despres utilitzar aquest token a la part on posa Authorize. Per\n  mes informacio, obtenir un usuari o altres peticions utilitzar el link de Contact Equip Sol o mandar un mail a jaumefullana@paucasesnovescifp.cat",
        "contact": {
            "name": "Equip Sol",
            "email": "jaumefullana@paucasesnovescifp.cat"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:81/allotjamentssol/public/"
        }
    ],
    "paths": {
...
        "/api/comentari": {
            "get": {
                "tags": [
                    "Comentaris"
                ],
                "summary": "Mostrar tots els comentaris.",
                "operationId": "51606adb66776f1e2b3f9f14e28ff926",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "Result": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Comentaris"
                ],
                "summary": "Modifica un comentari",
                "description": "Modifica un comentari.",
                "operationId": "0a7d50a61bae7dca1706d7eed56776f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    },
                                    "Comentari": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Aquest allotjament no es per tant"
                                    },
                                    "Valoracio": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "Result": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "Result": {
                                            "type": "string",
                                            "example": "Informacio de l'error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comentaris"
                ],
                "summary": "Crea un nou comentari",
                "description": "Crea un nou comentari.",
                "operationId": "998f855d6b60172566e09b5ad33bc708",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Comentari": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Aquest allotjament es un 10"
                                    },
                                    "Valoracio": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "5"
                                    },
                                    "DataCreacio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-03-21"
                                    },
                                    "AllotjamentsID ": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "Result": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "Result": {
                                            "type": "string",
                                            "example": "Informacio de l'error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comentaris"
                ],
                "summary": "Esborra un comentari",
                "description": "Esborra un comentari.",
                "operationId": "f0a9d6c4b91583082c2ecddb1c8dc57c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "Result": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "Result": {
                                            "type": "string",
                                            "example": "Informacio de l'error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comentari/{id}": {
            "get": {
                "tags": [
                    "Comentaris"
                ],
                "summary": "Mostrar el comentari al qual correspon la id passada per la url",
                "operationId": "d78adb10af484ae252284dd4348df414",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del comentari",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "Result": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}